Refactor weight calculation and lifeline handling: update weight answer return type, adjust bias logic in AnswerItem, and modify Lifeline percentage type

+++ app/Domain/Entities/ItemQuestionEntity.php
@@ -87,7 +87,7 @@ class ItemQuestionEntity
         $this->lifelines = $lifelines;
     }
 
-    public function getWeightAnswers(): float
+    public function getWeightAnswers(): int
     {
         return array_reduce(
             $this->answers,
